backend/requirements.txt
Django>=4.1
djangorestframework
gunicorn
psycopg2-binary
python-dotenv
